Attemp to abstractualize Storage and move it into its seperate folder, *unfinished*
