+++ lib/gemstash.rb
@@ -27,4 +27,7 @@ module Gemstash
   autoload :Web,                 "gemstash/web"
   autoload :WebError,            "gemstash/http_client"
   autoload :VERSION,             "gemstash/version"
+  class Storage # :nodoc:
+   autoload :LocalService, "gemstash/storage/local_storage"
+  end
 end
