+++ lib/gemstash/configuration.rb
@@ -20,7 +20,8 @@ module Gemstash
       puma_threads: 16,
       puma_workers: 1,
       cache_expiration: 30 * 60,
-      cache_max_size: 500
+      cache_max_size: 500,
+      storage_service: "local"
     }.freeze
 
     DEFAULT_FILE = File.expand_path("~/.gemstash/config.yml").freeze
